🎨 apply prettier formatting to moved files
