+++ packages/browser-core/src/browser/pageMayExitObservable.ts
@@ -45,5 +45,3 @@ export function createPageMayExitObservable(): Observable<PageMayExitEvent> {
     }
   })
 }
-
-
