+++ packages/browser-core/src/index.ts
@@ -71,7 +71,6 @@ export {
   bridgeSupports,
   BridgeCapability,
   createBatch,
-  createFlushController,
   FLUSH_DURATION_LIMIT,
 } from './transport'
 export * from './tools/display'
