+++ packages/browser-core/src/transport/flushController.ts
@@ -1,2 +1,2 @@
-export { PageExitReason, createFlushController, FLUSH_DURATION_LIMIT, MESSAGES_LIMIT } from '@datadog/js-core/transport'
-export type { PageMayExitEvent, UrgentFlushReason, FlushReason, FlushEvent, FlushController } from '@datadog/js-core/transport'
+export { PageExitReason, FLUSH_DURATION_LIMIT } from '@datadog/js-core/transport'
+export type { PageMayExitEvent, UrgentFlushReason, FlushReason, FlushEvent } from '@datadog/js-core/transport'
