+++ packages/browser-core/test/index.ts
@@ -13,7 +13,6 @@ export * from './emulate/mockSyntheticsWorkerValues'
 export * from './emulate/mockVisibilityState'
 export * from './emulate/mockNavigator'
 export * from './emulate/mockEventBridge'
-export * from './emulate/mockFlushController'
 export * from './emulate/mockFetch'
 export * from './emulate/mockXhr'
 export * from './emulate/mockEventTarget'
