+++ packages/js-core/src/transport/flushController.spec.ts
@@ -1,10 +1,11 @@
-import { RECOMMENDED_REQUEST_BYTES_LIMIT } from '@datadog/js-core/transport'
-import type { Clock } from '../../test'
-import { mockClock } from '../../test'
-import type { PageMayExitEvent } from '../browser/pageMayExitObservable'
-import { Observable } from '../tools/observable'
-import type { FlushController, FlushEvent } from './flushController'
+import type { Clock } from '../../test/mockClock'
+import { mockClock } from '../../test/mockClock'
+import { Observable } from '../util/observable'
+import { RECOMMENDED_REQUEST_BYTES_LIMIT } from './payload'
+import type { PageMayExitEvent, FlushEvent } from './flushController'
 import { createFlushController, FLUSH_DURATION_LIMIT, MESSAGES_LIMIT } from './flushController'
+
+type FlushController = ReturnType<typeof createFlushController>
 
 const BYTES_LIMIT = RECOMMENDED_REQUEST_BYTES_LIMIT
 // Arbitrary message size that is below the BYTES_LIMIT
